+++ R/PlackettLuce.R
@@ -2,7 +2,7 @@
 #'
 #' Fit a Plackett-Luce model to a set of rankings. The rankings may be partial
 #' (each ranking completely ranks a subset of the items) and include ties of
-#' any order.
+#' arbitrary order.
 #'
 #' @section Model definition:
 #'
@@ -163,6 +163,17 @@
 #' \code{\link{pudding}}, \code{\link{preflib}}.
 #'
 #' Vignette: \code{vignette("Overview", package = "PlackettLuce")}.
+#'
+#' @note As the maximum tie order increases, the number of possible choices for
+#' each rank increases rapidly, particularly when the total number of items is
+#' high. This means that the model will be slower to fit with higher \eqn{D}.
+#' Moreover, due to the current implementation of the `vcov()` method,
+#' computation of the standard errors (as by `summary()`) becomes extremely
+#' slow or infeasible due to memory limits. Therefore we recommend
+#' `PlackettLuce()` for ties up to order 4. For higher order ties, a
+#' rank-ordered logit model, see [ROlogit::rologit()] or
+#' generalized Mallows Model as in [BayesMallows::compute_mallows()] may be
+#' more suitable.
 #'
 #' @param rankings a \code{"\link{rankings}"} object, or an object that can be
 #' coerced by \code{as.rankings}.  An [`"aggregated_rankings"`][aggregate()]